# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NeuroDesk AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type ChatbotStatus = typeof ChatbotStatus[keyof typeof ChatbotStatus];


export const ChatbotStatus = {
  draft: 'draft',
  training: 'training',
  ready: 'ready',
  error: 'error',
} as const;

export type ChatbotPersonality = typeof ChatbotPersonality[keyof typeof ChatbotPersonality];


export const ChatbotPersonality = {
  professional: 'professional',
  friendly: 'friendly',
  sales: 'sales',
  technical: 'technical',
  luxury: 'luxury',
} as const;

export interface Chatbot {
  id: number;
  userId: string;
  name: string;
  companyName: string;
  /** @nullable */
  description?: string | null;
  status: ChatbotStatus;
  personality: ChatbotPersonality;
  primaryColor: string;
  welcomeMessage: string;
  /** @nullable */
  trainingScore?: number | null;
  totalMessages: number;
  totalConversations: number;
  createdAt: string;
  updatedAt: string;
}

export type ChatbotInputPersonality = typeof ChatbotInputPersonality[keyof typeof ChatbotInputPersonality];


export const ChatbotInputPersonality = {
  professional: 'professional',
  friendly: 'friendly',
  sales: 'sales',
  technical: 'technical',
  luxury: 'luxury',
} as const;

export interface ChatbotInput {
  name: string;
  companyName: string;
  description?: string;
  personality?: ChatbotInputPersonality;
  primaryColor?: string;
  welcomeMessage?: string;
}

export type ChatbotUpdatePersonality = typeof ChatbotUpdatePersonality[keyof typeof ChatbotUpdatePersonality];


export const ChatbotUpdatePersonality = {
  professional: 'professional',
  friendly: 'friendly',
  sales: 'sales',
  technical: 'technical',
  luxury: 'luxury',
} as const;

export interface ChatbotUpdate {
  name?: string;
  companyName?: string;
  description?: string;
  personality?: ChatbotUpdatePersonality;
  primaryColor?: string;
  welcomeMessage?: string;
}

export interface EmbedCode {
  snippet: string;
  botId: number;
}

export type KnowledgeSourceType = typeof KnowledgeSourceType[keyof typeof KnowledgeSourceType];


export const KnowledgeSourceType = {
  url: 'url',
  text: 'text',
  faq: 'faq',
} as const;

export type KnowledgeSourceStatus = typeof KnowledgeSourceStatus[keyof typeof KnowledgeSourceStatus];


export const KnowledgeSourceStatus = {
  pending: 'pending',
  processed: 'processed',
  error: 'error',
} as const;

export interface KnowledgeSource {
  id: number;
  chatbotId: number;
  type: KnowledgeSourceType;
  title: string;
  content: string;
  status: KnowledgeSourceStatus;
  createdAt: string;
}

export type KnowledgeSourceInputType = typeof KnowledgeSourceInputType[keyof typeof KnowledgeSourceInputType];


export const KnowledgeSourceInputType = {
  url: 'url',
  text: 'text',
  faq: 'faq',
} as const;

export interface KnowledgeSourceInput {
  type: KnowledgeSourceInputType;
  title: string;
  content: string;
}

export interface ChatMessageInput {
  message: string;
  sessionId: string;
}

export interface BotConversation {
  id: number;
  chatbotId: number;
  sessionId: string;
  messageCount: number;
  createdAt: string;
}

export type BotMessageRole = typeof BotMessageRole[keyof typeof BotMessageRole];


export const BotMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface BotMessage {
  id: number;
  conversationId: number;
  role: BotMessageRole;
  content: string;
  createdAt: string;
}

export interface BotConversationWithMessages {
  id: number;
  chatbotId: number;
  sessionId: string;
  messages: BotMessage[];
  createdAt: string;
}

export interface TopBot {
  id: number;
  name: string;
  messageCount: number;
}

export interface DashboardStats {
  totalChatbots: number;
  totalConversations: number;
  totalMessages: number;
  activeChatbots: number;
  messagesThisMonth: number;
  topBots: TopBot[];
}

export interface ChatbotStats {
  totalConversations: number;
  totalMessages: number;
  averageMessagesPerConversation: number;
  /** @nullable */
  trainingScore?: number | null;
  sourceCount: number;
}

export interface Plan {
  id: string;
  name: string;
  priceMonthly: number;
  priceYearly: number;
  /** @nullable */
  chatbotLimit?: number | null;
  messageLimit: number;
  features: string[];
  popular: boolean;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}