# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NeuroDesk AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chatbots:
    get:
      operationId: listChatbots
      tags: [chatbots]
      summary: List all chatbots for the current user
      responses:
        "200":
          description: List of chatbots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chatbot"
    post:
      operationId: createChatbot
      tags: [chatbots]
      summary: Create a new chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotInput"
      responses:
        "201":
          description: Created chatbot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chatbot"
  /chatbots/{id}:
    get:
      operationId: getChatbot
      tags: [chatbots]
      summary: Get a chatbot by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chatbot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chatbot"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateChatbot
      tags: [chatbots]
      summary: Update a chatbot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotUpdate"
      responses:
        "200":
          description: Updated chatbot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chatbot"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteChatbot
      tags: [chatbots]
      summary: Delete a chatbot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /chatbots/{id}/train:
    post:
      operationId: trainChatbot
      tags: [chatbots]
      summary: Trigger training for a chatbot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Training started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chatbot"
  /chatbots/{id}/embed-code:
    get:
      operationId: getChatbotEmbedCode
      tags: [chatbots]
      summary: Get the embed script snippet for a chatbot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Embed code snippet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedCode"

  # ── Knowledge sources ─────────────────────────────────────
  /chatbots/{id}/sources:
    get:
      operationId: listSources
      tags: [knowledge]
      summary: List all knowledge sources for a chatbot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeSource"
    post:
      operationId: createSource
      tags: [knowledge]
      summary: Add a knowledge source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeSourceInput"
      responses:
        "201":
          description: Created source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeSource"
  /chatbots/{id}/sources/{sourceId}:
    delete:
      operationId: deleteSource
      tags: [knowledge]
      summary: Delete a knowledge source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sourceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Widget conversations ───────────────────────────────────
  /chatbots/{id}/chat:
    post:
      operationId: chatWithBot
      tags: [conversations]
      summary: Send a message to a chatbot widget (streaming SSE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: SSE stream of AI response chunks
          content:
            text/event-stream: {}
  /chatbots/{id}/conversations:
    get:
      operationId: listBotConversations
      tags: [conversations]
      summary: List all widget conversations for a chatbot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BotConversation"
  /chatbots/{id}/conversations/{convId}:
    get:
      operationId: getBotConversation
      tags: [conversations]
      summary: Get a conversation with its messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: convId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConversationWithMessages"

  # ── Stats / analytics ─────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Chatbots ──────────────────────────────────────────────
    Chatbot:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        companyName:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, training, ready, error]
        personality:
          type: string
          enum: [professional, friendly, sales, technical, luxury]
        primaryColor:
          type: string
        welcomeMessage:
          type: string
        trainingScore:
          type: ["integer", "null"]
        totalMessages:
          type: integer
        totalConversations:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - companyName
        - status
        - personality
        - primaryColor
        - welcomeMessage
        - totalMessages
        - totalConversations
        - createdAt
        - updatedAt
    ChatbotInput:
      type: object
      properties:
        name:
          type: string
        companyName:
          type: string
        description:
          type: string
        personality:
          type: string
          enum: [professional, friendly, sales, technical, luxury]
        primaryColor:
          type: string
        welcomeMessage:
          type: string
      required:
        - name
        - companyName
    ChatbotUpdate:
      type: object
      properties:
        name:
          type: string
        companyName:
          type: string
        description:
          type: string
        personality:
          type: string
          enum: [professional, friendly, sales, technical, luxury]
        primaryColor:
          type: string
        welcomeMessage:
          type: string
    EmbedCode:
      type: object
      properties:
        snippet:
          type: string
        botId:
          type: integer
      required:
        - snippet
        - botId

    # ── Knowledge sources ─────────────────────────────────────
    KnowledgeSource:
      type: object
      properties:
        id:
          type: integer
        chatbotId:
          type: integer
        type:
          type: string
          enum: [url, text, faq]
        title:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [pending, processed, error]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - chatbotId
        - type
        - title
        - content
        - status
        - createdAt
    KnowledgeSourceInput:
      type: object
      properties:
        type:
          type: string
          enum: [url, text, faq]
        title:
          type: string
        content:
          type: string
      required:
        - type
        - title
        - content

    # ── Widget conversations ───────────────────────────────────
    ChatMessageInput:
      type: object
      properties:
        message:
          type: string
        sessionId:
          type: string
      required:
        - message
        - sessionId
    BotConversation:
      type: object
      properties:
        id:
          type: integer
        chatbotId:
          type: integer
        sessionId:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - chatbotId
        - sessionId
        - messageCount
        - createdAt
    BotConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        chatbotId:
          type: integer
        sessionId:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/BotMessage"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - chatbotId
        - sessionId
        - messages
        - createdAt
    BotMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    # ── Stats ─────────────────────────────────────────────────
...