# lib/api-zod/src/generated/types/chatbotInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NeuroDesk AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatbotInputPersonality } from './chatbotInputPersonality';

export interface ChatbotInput {
  name: string;
  companyName: string;
  description?: string;
  personality?: ChatbotInputPersonality;
  primaryColor?: string;
  welcomeMessage?: string;
}