# lib/api-zod/src/generated/types/chatbotUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NeuroDesk AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatbotUpdatePersonality } from './chatbotUpdatePersonality';

export interface ChatbotUpdate {
  name?: string;
  companyName?: string;
  description?: string;
  personality?: ChatbotUpdatePersonality;
  primaryColor?: string;
  welcomeMessage?: string;
}