# lib/api-zod/src/generated/types/embedCode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NeuroDesk AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmbedCode {
  snippet: string;
  botId: number;
}