# lib/api-zod/src/generated/types/knowledgeSourceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NeuroDesk AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KnowledgeSourceInputType } from './knowledgeSourceInputType';

export interface KnowledgeSourceInput {
  type: KnowledgeSourceInputType;
  title: string;
  content: string;
}